Add a very dumb example fuzzer for SQLite.
It doesn't seem a very useful fuzzer, because it almost never compiles
a working statement. It was intended as an example of how you might use
the dictionary, but does not seem useful.

It uses a dictionary of words that might be valid for passing to the
SQLite engine, but the statements it generates don't seem to ever go
anywhere useful - the coverage sits at around 13 even after thousands
of executions. It's possible this is because the bulk of the work is
in the C-side so we will never see progress. Plus, SQLite is pretty
solid.